added a new file per type
