+++ src/components/content/card.tsx
@@ -2,7 +2,7 @@ import { CARD_LAYOUT } from "@/utils/constants/contentful";
 import ImageComponent from "../common/image-component";
 import LinkComponent from "../common/link-component";
 import Description from "../common/description";
-import { CardType } from "@/domain/cms/components/card";
+import { CardType } from "@/domain/components/card-type";
 
 interface CardParams {
   readonly data: CardType;
