--- src/domain/cms/components/content-block.ts
@@ -1,18 +1,0 @@
-import { ImageResponse, ImageType } from "./asset";
-import { LinkComponentType } from "./link-component";
-
-export type ContentBlockResponse = {
-  title?: string;
-  description?: {
-    json: Node;
-  };
-  image?: ImageResponse;
-  link?: LinkComponentType;
-};
-
-export type ContentBlockType = {
-  title?: string;
-  description?: string;
-  image?: ImageType;
-  link?: LinkComponentType;
-};
