+++ src/domain/components/card-block-type.ts
@@ -1,0 +1,7 @@
+import { CardType } from "./card-type";
+
+export type CardBlockType = {
+  title?: string;
+  layout: "1 - 1 - 1" | "1 - 2" | "2 - 1" | "1 - 1 - 1 - 1";
+  cards: CardType[];
+};
