+++ src/domain/components/carousel-type.ts
@@ -1,0 +1,6 @@
+import { ImageType } from "./image-type";
+
+export type CarouselType = {
+  title?: string;
+  carouselItems: ImageType[];
+};
