+++ src/domain/components/content-block-type.ts
@@ -1,0 +1,9 @@
+import { ImageType } from "./image-type";
+import { LinkComponentType } from "./link-component-type";
+
+export type ContentBlockType = {
+  title?: string;
+  description?: string;
+  image?: ImageType;
+  link?: LinkComponentType;
+};
