+++ src/domain/components/image-type.ts
@@ -1,0 +1,7 @@
+export type ImageType = {
+  id: string;
+  title: string;
+  url: string;
+  height: number;
+  width: number;
+};
