+++ src/domain/components/link-component-type.ts
