+++ src/domain/misc/collection.ts
@@ -1,0 +1,3 @@
+export type Collection<T> = {
+  items: T[];
+};
