+++ src/domain/misc/get-content-by-page-params.ts
@@ -1,7 +1,3 @@
-export type Collection<T> = {
-  items: T[];
-};
-
 export type GetContentByPageParams = {
   page: string;
   locale?: string;
