+++ src/domain/pages/static-page-type.ts
@@ -1,0 +1,6 @@
+import { ContentBlockType } from "../components/content-block-type";
+
+export type StaticPageType = {
+  title: string;
+  content: ContentBlockType[];
+};
