+++ src/domain/response/cms/card-block-response.ts
@@ -1,0 +1,8 @@
+import { Collection } from "@/domain/common";
+import { CardResponse } from "./card-response";
+
+export type CardBlockResponse = {
+  title?: string;
+  layout: "1 - 1 - 1" | "1 - 2" | "2 - 1" | "1 - 1 - 1 - 1";
+  cardsCollection: Collection<CardResponse>;
+};
