+++ src/domain/response/cms/carousel-response.ts
@@ -1,0 +1,7 @@
+import { Collection } from "@/domain/misc/collection";
+import { ImageResponse } from "./image-response";
+
+export type CarouselResponse = {
+  title?: string;
+  carouselItemsCollection: Collection<ImageResponse>;
+};
