+++ src/domain/response/cms/homepage-collection-response.ts
@@ -1,0 +1,6 @@
+import { Collection } from "@/domain/misc/collection";
+import { HomepageResponse } from "./homepage-response";
+
+export type HomepageCollectionResponse = {
+  homepageCollection: Collection<HomepageResponse>;
+};
