+++ src/domain/response/cms/homepage-response.ts
@@ -1,0 +1,11 @@
+import { CardBlockResponse } from "./card-block-response";
+import { CarouselResponse } from "./carousel-response";
+
+export type HomepageResponse = {
+  title: string;
+  description: {
+    json: Node;
+  };
+  carousel?: CarouselResponse;
+  cardBlock: CardBlockResponse;
+};
