+++ src/domain/response/cms/image-response.ts
@@ -7,11 +7,3 @@ export type ImageResponse = {
   height: number;
   width: number;
 };
-
-export type ImageType = {
-  id: string;
-  title: string;
-  url: string;
-  height: number;
-  width: number;
-};
