+++ src/domain/response/cms/static-page-collection-response.ts
@@ -1,0 +1,6 @@
+import { Collection } from "@/domain/misc/collection";
+import { StaticPageResponse } from "./static-page-response";
+
+export type StaticPageCollectionResponse = {
+  staticPageCollection: Collection<StaticPageResponse>;
+};
