+++ src/domain/response/cms/static-page-response.ts
@@ -1,0 +1,7 @@
+import { Collection } from "@/domain/misc/collection";
+import { ContentBlockResponse } from "./content-block-response";
+
+export type StaticPageResponse = {
+  title: string;
+  contentCollection: Collection<ContentBlockResponse>;
+};
